refactor(connlib): simplify resource ordering by implementing Ord trait
